refactor: MovingCarStrategy의 구현 클래스인 RandomMovingStrategy 클래스 생성
