--- src/main/java/racingcar/service/RandomNumberService.java
@@ -1,8 +1,0 @@
-package racingcar.service;
-
-public class RandomNumberService {
-	public static int getRandomNumber() {
-		int random = (int)(Math.random() * 10);
-		return random;
-	}
-}
