+++ src/main/java/racingcar/strategy/RandomMovingStrategy.java
@@ -1,0 +1,14 @@
+package racingcar.strategy;
+
+public class RandomMovingStrategy implements MovingCarStrategy {
+    private static final int MOVING_CONDITION_NUMBER = 4;
+
+    @Override
+    public boolean canMove() {
+        return getRandomNumber() > MOVING_CONDITION_NUMBER;
+    }
+
+    private int getRandomNumber() {
+        return (int) (Math.random() * 10);
+    }
+}
